Fixed bug in composite queries and added global for dropping tables in cassandra

+++ AppDB/cassandra/cassandra_helper.rb
@@ -2,6 +2,8 @@ require 'djinn'
 require 'djinn_job_data'
 require 'helperfunctions'
 
+DROP_TABLES = true
+
 def get_uaserver_ip()
   Djinn.get_nearest_db_ip
 end
@@ -62,8 +64,11 @@ def start_db_master()
   Djinn.log_debug("Starting up Cassandra as master")
 
   Djinn.log_run("pkill ThriftBroker")
-  `rm -rf /var/appscale/cassandra*`
-  `rm /var/log/appscale/cassandra/system.log`
+  if DROP_TABLES
+    `rm -rf /var/appscale/cassandra*`
+    `rm /var/log/appscale/cassandra/system.log`
+  end
+  
   Djinn.log_run("#{APPSCALE_HOME}/AppDB/cassandra/cassandra/bin/cassandra start -p /var/appscale/appscale-cassandra.pid")
   HelperFunctions.sleep_until_port_is_open(HelperFunctions.local_ip, 9160)
 end
@@ -74,10 +79,12 @@ def start_db_slave()
 
   HelperFunctions.sleep_until_port_is_open(Djinn.get_db_master_ip, 9160)
   sleep(5)
-  `rm -rf /var/appscale/cassandra*`
-  `rm /var/log/appscale/cassandra/system.log`
+  if DROP_TABLES
+    `rm -rf /var/appscale/cassandra*`
+    `rm /var/log/appscale/cassandra/system.log`
+  end
   `#{APPSCALE_HOME}/AppDB/cassandra/cassandra/bin/cassandra start -p /var/appscale/appscale-cassandra.pid`
-  #Djinn.log_run("#{APPSCALE_HOME}/AppDB/cassandra/cassandra/bin/cassandra start -p /var/appscale/appscale-cassandra.pid")
+  Djinn.log_run("#{APPSCALE_HOME}/AppDB/cassandra/cassandra/bin/cassandra start -p /var/appscale/appscale-cassandra.pid")
   HelperFunctions.sleep_until_port_is_open(HelperFunctions.local_ip, 9160)
 end
 
@@ -88,7 +95,9 @@ end
 
 def stop_db_slave
   Djinn.log_debug("Stopping Cassandra slave")
-  Djinn.log_run("#{APPSCALE_HOME}/AppDB/cassandra/cassandra/bin/nodetool decommission -h #{HelperFunctions.local_ip} -p 6666")
+  if DROP_TABLES
+    Djinn.log_run("#{APPSCALE_HOME}/AppDB/cassandra/cassandra/bin/nodetool decommission -h #{HelperFunctions.local_ip} -p 6666")
+  end
   Djinn.log_run("cat /var/appscale/appscale-cassandra.pid | xargs kill -9")
 end
